Move IPC logic to its own module
